2009-01-12  Sancho Lerena <lsrena@artica.es>
        * functions_db.php,
        * godmode/agentes/configurar_agente.php: Delete module deletes
        tagent_data_inc record (it's a very small table).

        * reporting/fgraph.php: Fixed small problem in tagent_access graph.

        * pandoradb_migrate_20_to_21.sql: DB Schema migration script. Not
        finished (not convert old data, only add/remove structures).



git-svn-id: https://svn.code.sf.net/p/pandora/code/trunk@1338 c3f86ba8-e40f-0410-aaad-9ba5e7f4b01f

+++ pandora_console/include/functions_db.php
@@ -2169,18 +2169,23 @@ function delete_agent ($id_agents) {
 		//Access entries
 		temp_sql_delete ("tagent_access", "id_agent", $id_agent);
 
+		//tagente_datos_inc
+                temp_sql_delete ("tagente_datos_inc", "id_agente_modulo", $tmodbase);
+
 		//And at long last, the agent
 		temp_sql_delete ("tagente", "id_agente", $id_agent);
 		
 		// Delete remote configuration
 		$agent_md5 = md5($agent_name, FALSE);
-		if (file_exists($config["remote_config"] . "/" . $agent_md5 . ".md5")) {
-		// Agent remote configuration editor
-			$file_name = $config["remote_config"] . "/" . $agent_md5 . ".conf";
-			unlink ($file_name);
-			$file_name = $config["remote_config"] . "/" . $agent_md5 . ".md5";
-			unlink ($file_name);
-		}
+
+		if (isset($config["remote_config"]))
+			if (file_exists($config["remote_config"] . "/" . $agent_md5 . ".md5")) {
+				// Agent remote configuration editor
+				$file_name = $config["remote_config"] . "/" . $agent_md5 . ".conf";
+				unlink ($file_name);
+				$file_name = $config["remote_config"] . "/" . $agent_md5 . ".md5";
+				unlink ($file_name);
+			}
 	}
 
 	if ($errors > 0) {
